+++ source/jormungandr/jormungandr/interfaces/v1/Places.py
@@ -46,7 +46,6 @@ from jormungandr.exceptions import TechnicalError
 from functools import wraps
 from flask_restful import marshal, marshal_with
 import datetime
-from jormungandr.autocomplete.elastic_search import Elasticsearch
 from jormungandr.parking_space_availability.bss.stands_manager import ManageStands
